api/requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
pillow